Handle different ID perspectives
The same event has a different ID from the perspective of the organizer
and from the room. This commit makes it unambiguous that we always store
the ID from the resource's perspective, and we include the UID.

The ChangeKey is no longer relevant since changing the approach for
recognizing changes.